redircting users instead of showing banner
