+++ src/common/utils/utility.js
@@ -1,3 +1,6 @@
+import { generateLiveApiInstance } from '../../common/appId';
+import { getTokenList } from '../../common/utils/storageManager';
+
 const euCountries = [
     'it',
     'de',
@@ -32,3 +35,29 @@ const euCountries = [
 export const isEuCountry = country => euCountries.includes(country);
 
 export const isUKCountry = country => country === 'gb';
+
+/* eslint-disable camelcase */
+export const moveToDeriv = async () => {
+    const api = generateLiveApiInstance();
+    console.log('api:', api);
+    const { website_status } = await api.send({ website_status: 1 });
+    const { clients_country } = website_status;
+    const tokenList = getTokenList();
+    const landingCompanyName = tokenList.map(token => token.loginInfo.landing_company_name);
+
+    if (!tokenList.length) {
+        if (isEuCountry(clients_country) || isUKCountry(clients_country)) {
+            window.location.replace('https://binary.com/move-to-deriv');
+        }
+    }
+
+    if (
+        (landingCompanyName.length === 1 &&
+            landingCompanyName.includes('virtual') &&
+            (isEuCountry(clients_country) || isUKCountry(clients_country))) ||
+        landingCompanyName.includes('maltainvest') ||
+            landingCompanyName.includes('malta') || landingCompanyName.includes('iom')
+    ) {
+        window.location.replace('https://binary.com/move-to-deriv');
+    }
+};
