+++ src/indexPage/index.js
@@ -7,13 +7,12 @@ import { oauthLogin } from '../common/appId';
 import { isEuCountry, showHideEuElements } from '../common/footer-checks';
 import GTM from '../common/gtm';
 import { load as loadLang, showBanner } from '../common/lang';
+import { moveToDeriv } from '../common/utils/utility';
 import { getTokenList } from '../common/utils/storageManager';
 import { createUrl } from '../common/utils/tools';
 import '../common/binary-ui/dropdown';
-import NotificationBanner from './react-components/NotificationBanner.jsx';
 
 const renderElements = () => {
-    ReactDOM.render(<NotificationBanner />, document.getElementById('notification-banner'));
     ReactDOM.render(<Logo />, document.getElementById('binary-logo'));
     ReactDOM.render(<Footer />, document.getElementById('footer'));
     isEuCountry().then(isEu => showHideEuElements(isEu));
@@ -24,6 +23,7 @@ const renderElements = () => {
 
 const loginCheck = () => {
     if (endpoint()) return;
+    moveToDeriv();
     if (getTokenList().length) {
         window.location.pathname = `${window.location.pathname.replace(/\/+$/, '')}/bot.html`;
     } else {
